This makes the tic tac toe "super"
The point of this branch is to automatize the super tic tac toe board,
meaning the game could be played at any level. With this commit, the
user can obtain a board of any level, and play x's and circles in any
square. The user can't, however, actually play the game. The next step
is to rework the game mechanics to be as automatic as the board.